remove unnecessary files and update deployment scripts
