+++ scripts/dip38-1.ts
@@ -64,10 +64,7 @@ const handler = async (
 	// Create the LockupMigration
 	const lockup = new Lockup(dev)
 	const lockup_current = await lockup.load()
-	const lockup_migration = new LockupMigration(dev)
-	const lockup_next = await lockup_migration.create(
-		await lockup_current.devMinter()
-	)
+	const lockup_next = await lockup.create(await lockup_current.devMinter())
 
 	// Create the new Withdraw
 	const withdraw = new Withdraw(dev)
@@ -79,10 +76,7 @@ const handler = async (
 	// Delegate to all new contracts
 	await Promise.all([
 		metrics_group.changeOwner(metrics_group_current, metrics_group_next),
-		lockup_migration.changeOwnerToMigrationContract(
-			lockup_current,
-			lockup_next
-		),
+		lockup.changeOwner(lockup_current, lockup_next),
 		withdraw.changeOwner(withdraw_current, withdraw_next),
 	])
